rename images package image
